remove some vg functions
