+++ src/genotyper.cpp
@@ -2,6 +2,9 @@
 #include "genotyper.hpp"
 #include "traversal_finder.hpp"
 #include "cactus_snarl_finder.hpp"
+#include "path_index.hpp"
+#include "utility.hpp"
+#include "translator.hpp"
 
 //#define debug
