+++ src/genotyper.hpp
@@ -16,14 +16,9 @@
 #include <list>
 #include <vg/vg.pb.h>
 #include "vg.hpp"
-#include "translator.hpp"
-#include "srpe.hpp"
 #include "hash_map.hpp"
-#include "utility.hpp"
 #include "types.hpp"
 #include "genotypekit.hpp"
-#include "srpe.hpp"
-#include "path_index.hpp"
 #include "statistics.hpp"
 
 namespace vg {
