+++ src/unittest/chunker.cpp
@@ -4,6 +4,8 @@
 
 #include "catch.hpp"
 #include "chunker.hpp"
+#include "vg.hpp"
+#include "xg.hpp"
 #include "path.hpp"
 
 namespace vg {
