+++ src/unittest/handle.cpp
@@ -669,15 +669,18 @@ TEST_CASE("divide_handle and destroy_handle work together on both strands", "[ha
             REQUIRE(steps.size() == 9);
             int i = 0;
             vector<handle_t> to_delete;
-            g->append_step(g->create_path_handle(to_string(i)), steps[i++]);            
+            g->append_step(g->create_path_handle(to_string(i)), steps[i]);
             to_delete.push_back(steps[i++]);
-            g->append_step(g->create_path_handle(to_string(i)), steps[i++]);
+            g->append_step(g->create_path_handle(to_string(i)), steps[i]);
+            ++i;
             to_delete.push_back(steps[i++]);
             to_delete.push_back(steps[i++]);
             to_delete.push_back(steps[i++]);
-            g->append_step(g->create_path_handle(to_string(i)), steps[i++]);
+            g->append_step(g->create_path_handle(to_string(i)), steps[i]);
+            ++i;
             to_delete.push_back(steps[i++]);
-            g->append_step(g->create_path_handle(to_string(i)), steps[i++]);            
+            g->append_step(g->create_path_handle(to_string(i)), steps[i]);
+            ++i;
                                 
             g->destroy_path(path_handle);
