+++ src/unittest/sampler.cpp
@@ -10,6 +10,7 @@
 #include <vg/vg.pb.h>
 #include "../sampler.hpp"
 #include "../xg.hpp"
+#include "../vg.hpp"
 #include "catch.hpp"
 
 namespace vg {
