+++ src/variant_adder.hpp
@@ -8,6 +8,7 @@
 
 #include "vcf_buffer.hpp"
 #include "path_index.hpp"
+#include "handle.hpp"
 #include "vg.hpp"
 #include "xg.hpp"
 #include "progressive.hpp"
