+++ src/variant_recall.hpp
@@ -14,17 +14,10 @@
 #include <list>
 #include <vg/vg.pb.h>
 #include "vg.hpp"
-#include "xg.hpp"
-#include "translator.hpp"
 #include "deconstructor.hpp"
-#include "srpe.hpp"
 #include "hash_map.hpp"
-#include "utility.hpp"
 #include "types.hpp"
 #include "genotypekit.hpp"
-#include "srpe.hpp"
-#include "path_index.hpp"
-#include "statistics.hpp"
 
 namespace vg {
