+++ src/vg.cpp
@@ -1837,47 +1837,6 @@ bool VG::mapping_is_total_match(const Mapping& m) {
         && mapping_from_length(m) == get_node(m.position().node_id())->sequence().size();
 }
 
-Node* VG::merge_nodes(const list<Node*>& nodes) {
-    // make the new node (use the first one in the list)
-    assert(!nodes.empty());
-    Node* n = nodes.front();
-    nid_t nid = n->id();
-    // create edges to the node
-    for (auto& m : nodes) {
-        if (m != n) { // skip first, which we're using
-            //set<NodeSide> sides_of(NodeSide side);
-            nid_t id = m->id();
-            for (auto& s : sides_to(NodeSide(id, false))) {
-                create_edge(s, NodeSide(nid, false));
-            }
-            for (auto& s : sides_to(NodeSide(id, true))) {
-                create_edge(s, NodeSide(nid, true));
-            }
-            for (auto& s : sides_from(NodeSide(id, false))) {
-                create_edge(NodeSide(nid, false), s);
-            }
-            for (auto& s : sides_from(NodeSide(id, true))) {
-                create_edge(NodeSide(nid, true), s);
-            }
-        }
-    }
-    // reassign mappings in paths to the new node
-    hash_map<nid_t, nid_t> id_mapping;
-    for (auto& m : nodes) {
-        if (m != n) {
-            id_mapping[m->id()] = nid;
-        }
-    }
-    paths.swap_node_ids(id_mapping);
-    // and erase the old nodes
-    for (auto& m : nodes) {
-        if (m != n) {
-            destroy_node(m);
-        }
-    }
-    // return the node we merged into
-    return n;
-}
 
 void VG::build_node_indexes_no_init_size(void) {
     for (nid_t i = 0; i < graph.node_size(); ++i) {
@@ -2471,52 +2430,6 @@ map<nid_t, vcflib::Variant> VG::get_node_nid_to_variant(vcflib::VariantCallFile 
     return ret;
 }
 
-void VG::dice_nodes(int max_node_size) {
-    // We're going to chop up everything, so clear out the path ranks.
-    paths.clear_mapping_ranks();
-
-    if (max_node_size) {
-        vector<Node*> nodes; nodes.reserve(size());
-        for_each_node(
-            [this, &nodes](Node* n) {
-                nodes.push_back(n);
-            });
-        auto lambda =
-            [this, max_node_size](Node* n) {
-            int node_size = n->sequence().size();
-            if (node_size > max_node_size) {
-                int div = 2;
-                while (node_size/div > max_node_size) {
-                    ++div;
-                }
-                int segment_size = node_size/div;
-
-                // Make up all the positions to divide at
-                vector<int> divisions;
-                int last_division = 0;
-                while(last_division + segment_size < node_size) {
-                    // We can fit another division point
-                    last_division += segment_size;
-                    divisions.push_back(last_division);
-                }
-
-                // What segments are we making?
-                vector<Node*> segments;
-
-                // Do the actual division
-                divide_node(n, divisions, segments);
-            }
-        };
-        for (int i = 0; i < nodes.size(); ++i) {
-            lambda(nodes[i]);
-        }
-    }
-
-    // Set the ranks again
-    paths.rebuild_mapping_aux();
-    paths.compact_ranks();
-}
-
 string VG::trav_sequence(const NodeTraversal& trav) {
     string seq = trav.node->sequence();
     if (trav.backward) {
@@ -2525,252 +2438,7 @@ string VG::trav_sequence(const NodeTraversal& trav) {
         return seq;
     }
 }
-
-void VG::bluntify(void) {
-    // we bluntify the graph by converting it from an overlap graph,
-    // which is supported by the data format through the edge's overlap field,
-    // into a blunt-end string graph which algorithms in VG assume
-
-    // We manage this by turning nodes with overlaps into threads in a pinch
-    // graph, and pinching together on the overlaps.
-    
-    // TODO: this does not preserve existing paths. If your assembler spits out
-    // paths you want to preserve, you'll need to write code to do that here.
-    paths.clear();
-    
-    // First we have to validate the overlaps claimed by the edges.
-
-    // We populate this with edges with incorrect overlaps that we want to delete.
-    set<Edge*> bad_edges;
-    // Run in parallel as this can be very expensive
-    for_each_edge_parallel([&](Edge* edge) {
-        if (edge->overlap() > 0) {
-#ifdef debug
-            cerr << "claimed overlap " << edge->overlap() << endl;
-#endif
-            // derive and check the overlap seqs
-            auto from_seq = trav_sequence(NodeTraversal(get_node(edge->from()), edge->from_start()));
-            auto to_seq = trav_sequence(NodeTraversal(get_node(edge->to()), edge->to_end()));
-
-            if (edge->overlap() > from_seq.size()) edge->set_overlap(from_seq.size());
-            if (edge->overlap() > to_seq.size()) edge->set_overlap(to_seq.size());
-        }
-        });
-    
-    for (auto* edge : bad_edges) {
-        // Drop all the edges with incorrect overlaps
-        destroy_edge(edge);
-    }
-    
-    // Create a pinch graph.
-    // Note that in pinch graphs, orientation 1 = forward, and 0 = reverse
-    auto* pinch_graph = stPinchThreadSet_construct();
-    
-    // Have a function to add a node to the pinch graph if it's not there
-    // already, or get it if it is.
-    auto obtain_thread = [&](Node* node) {
-        // IF it's there, grab it
-        auto* found = stPinchThreadSet_getThread(pinch_graph, node->id());
-        if (found == nullptr) {
-            // If not, make it
-            found = stPinchThreadSet_addThread(pinch_graph, node->id(), 0, node->sequence().size());
-        }
-        return found;
-    }; 
-    
-    // We fill this with the nodes we're going to replace.
-    set<Node*> overlapped_nodes;
-    
-    // And this with the overlap edges which we remove from the graph (to be
-    // replaced by pinching parts of nodes).
-    vector<Edge*> overlap_edges; 
-    
-    for_each_edge([&](Edge* edge) {
-        // Now loop over only the real edges
-        if (edge->overlap() > 0) {
-            // and for each one with overlap
-            
-            // Make it a pair of NodeSides
-            NodeSide left;
-            NodeSide right;
-            tie(left, right) = NodeSide::pair_from_edge(*edge);
-            
-            // Grab the nodes
-            Node* left_node = get_node(left.node);
-            Node* right_node = get_node(right.node);
-            
-            // Mark them overlapped
-            overlapped_nodes.insert(left_node);
-            overlapped_nodes.insert(right_node);
-        
-            // Grab both the nodes' threads from the pinch graph
-            auto* left_thread = obtain_thread(left_node);
-            auto* right_thread = obtain_thread(right_node);
-            // And the nodes' lengths
-            size_t left_length = left_node->sequence().size();
-            size_t right_length = right_node->sequence().size();
-            
-            // What should be the first base merged on the left node? If we're
-            // on the start of the node, we start at 0, but if we're on the end
-            // of the node, we start however far in we need to be to allow for
-            // the overlap.
-            size_t left_start = left.is_end ? (left_length - edge->overlap()) : 0;
-            // And on the right node?
-            size_t right_start = right.is_end ? (right_length - edge->overlap()) : 0;
-            
-            // Make the actual pinch. We have to do opposite strands if we're
-            // pinching the same ends of nodes together. Otherwise the remaining
-            // pieces of the node will wind up on the same side of the resulting
-            // merged block. Note that we need to send true for a forward
-            // relative orientation.
-            stPinchThread_pinch(left_thread, right_thread, left_start, right_start, edge->overlap(), !(left.is_end == right.is_end));
-#ifdef debug
-            cerr << "pinched " << left_node->id() << ":" << left_start << " and "
-                << right_node->id() << ":" << right_start << " for "
-                << edge->overlap() << " bp in pinch orientation "
-                << !(left.is_end == right.is_end) << endl;
-#endif
-
-            // Remember to remove the edge from the graph
-            overlap_edges.push_back(edge);
-            
-        }
-    });
-    
-    // Remove all the overlap edges from the graph, since we represent them with
-    // stuff coming back from the pinch graph.
-    for (auto* edge : overlap_edges) {
-        destroy_edge(edge);
-    }
-    
-    // Clean up any trivial boundaries that might be in the pinch graph.
-    stPinchThreadSet_joinTrivialBoundaries(pinch_graph);
-    
-    // This maps from pinch block to new node created for it.
-    map<stPinchBlock*, Node*> new_block_nodes;
-    // Some segments aren't in blocks but still need nodes.
-    map<stPinchSegment*, Node*> new_segment_nodes;
-    
-    // We have a function that will get the node for a segment's block or for
-    // the segment, and the relative orientation of the node to the segment
-    // (true for reverse). It will create the node if it's not found.
-    auto obtain_node = [&](stPinchSegment* segment) {
-        auto* block = stPinchSegment_getBlock(segment);
-        
-        if (block == nullptr) {
-            // No block, just do node by segment
-            if (!new_segment_nodes.count(segment)) {
-                // Need to make the node. So grab the sequence.
-                Node* source_node = get_node(stPinchSegment_getName(segment));
-                string segment_sequence = source_node->sequence().substr(stPinchSegment_getStart(segment), stPinchSegment_getLength(segment));
-                new_segment_nodes[segment] = create_node(segment_sequence);
-            }
-            // Send it back saying it's definitely in the right orientation.
-            return make_pair(new_segment_nodes[segment], false);
-        } else {
-            // The segment is in a block, so we look for the block's node.
-            if (!new_block_nodes.count(block)) {
-                // Need to make the node. So grab the sequence.
-                Node* source_node = get_node(stPinchSegment_getName(segment));
-                string segment_sequence = source_node->sequence().substr(stPinchSegment_getStart(segment), stPinchSegment_getLength(segment));
-                if (!stPinchSegment_getBlockOrientation(segment)) {
-                    // This block is really the reverse complement of this segment
-                    segment_sequence = reverse_complement(segment_sequence);
-                }
-                new_block_nodes[block] = create_node(segment_sequence);
-            }
-            
-            // Send the node back, along with its VG-style orientation relative to the segment.
-            return make_pair(new_block_nodes[block], !stPinchSegment_getBlockOrientation(segment));
-        }
-    };
-    
-    // Now we go through the pinch segments
-    auto iterator = stPinchThreadSet_getSegmentIt(pinch_graph);
-    auto* segment = stPinchThreadSetSegmentIt_getNext(&iterator);
-    
-    while (segment != nullptr) {
-        // For each segment, we have to have a new node. Find it and its
-        // orientation relative to the segment (true = opposite orientations).
-        Node* segment_node;
-        bool segment_is_reverse;
-        tie(segment_node, segment_is_reverse) = obtain_node(segment);
-        
-        // What node did we break up?
-        Node* old_node = get_node(stPinchSegment_getName(segment));
-        
-#ifdef debug
-        cerr << "Handle segment " << segment << " of node " << old_node->id() << " as new node " << segment_node->id() << endl;
-#endif
-        
-        // We look at all the connected blocks, and if they have associated nodes, we create edges to them
-        
-        auto* prev_segment = stPinchSegment_get5Prime(segment);
-        if (prev_segment != nullptr) {
-            // There's more of this original node left of here
-            Node* prev_node;
-            bool prev_is_reverse;
-            tie(prev_node, prev_is_reverse) = obtain_node(prev_segment);
-            
-            // Make the edge (which may exist already)
-            create_edge(prev_node, segment_node, prev_is_reverse, segment_is_reverse);
-
-#ifdef debug
-            cerr << "\tAttach to prev segment " << prev_segment << " as new node " << prev_node->id() << endl;
-#endif
-            
-        } else {
-            // We're the start of our original node
-            NodeSide original_start(old_node->id(), false);
-            for (auto& attached : sides_of(original_start)) {
-                // For everything attached to the start of the original node, attach it to the left of here
-                create_edge(attached, NodeSide(segment_node->id(), segment_is_reverse));
-                
-#ifdef debug
-                cerr << "\tAttach to old side " << attached << " on the left" << endl;
-#endif
-            }
-        }
-        
-        auto* next_segment = stPinchSegment_get3Prime(segment);
-        if (next_segment != nullptr) {
-            // There's more of our original node right of here
-            Node* next_node;
-            bool next_is_reverse;
-            tie(next_node, next_is_reverse) = obtain_node(next_segment);
-            
-            // Make the edge (which may exist already)
-            create_edge(segment_node, next_node, segment_is_reverse, next_is_reverse);
-            
-#ifdef debug
-            cerr << "\tAttach to next segment " << next_segment << " as new node " << next_node->id() << endl;
-#endif
-            
-        } else {
-            // We're the end of our original node
-            NodeSide original_end(old_node->id(), true);
-            for (auto& attached : sides_of(original_end)) {
-                // For everything attached to the end of the original node, attach it to the right of here
-                create_edge(attached, NodeSide(segment_node->id(), !segment_is_reverse));
-                
-#ifdef debug
-                cerr << "\tAttach to old side " << attached << " on the right" << endl;
-#endif
-            }
-        }
-        
-        segment = stPinchThreadSetSegmentIt_getNext(&iterator);
-    }
-    
-    // Now clean up the pinch thread set. This invalidates all the pointers we got from it.
-    stPinchThreadSet_destruct(pinch_graph);
-    
-    for (auto* node : overlapped_nodes) {
-        // Then finally we go through all the nodes that had overlaps and remove them and their edges.
-        destroy_node(node);
-    }
-    
-}
+\
 
 static
 void
@@ -3085,88 +2753,6 @@ void VG::dfs(const function<void(NodeTraversal)>& node_begin_fn,
         NULL,
         NULL);
 }
-    
-// recursion-free version of Tarjan's strongly connected components algorithm
-// https://en.wikipedia.org/wiki/Tarjan%27s_strongly_connected_components_algorithm
-// Generalized to bidirected graphs as described (confusingly) in
-// "Decomposition of a bidirected graph into strongly connected components and
-// its signed poset structure", by Kazutoshi Ando, Satoru Fujishige, and Toshio
-// Nemoto. http://www.sciencedirect.com/science/article/pii/0166218X95000683
-
-// The best way to think about that paper is that the edges are vectors in a
-// vector space with number of dimensions equal to the number of nodes in the
-// graph, and an edge attaching to the end a node is the positive unit vector in
-// its dimension, and an edge attaching to the start of node is the negative
-// unit vector in its dimension.
-
-// The basic idea is that you just consider the orientations as different nodes,
-// and the edges as existing between both pairs of orientations they connect,
-// and do connected components on that graph. Since we don't care about
-// "consistent" or "inconsistent" strongly connected components, we just put a
-// node in a component if either orientation is in it. But bear in mind that
-// both orientations of a node might not actually be in the same strongly
-// connected component in a bidirected graph, so now the components may overlap.
-set<set<nid_t> > VG::strongly_connected_components(void) {
-    
-    // What node visit step are we on?
-    int64_t index = 0;
-    // What's the search root from which a node was reached?
-    map<NodeTraversal, NodeTraversal> roots;
-    // At what index step was each node discovered?
-    map<NodeTraversal, int64_t> discover_idx;
-    // We need our own copy of the DFS stack
-    deque<NodeTraversal> stack;
-    // And our own set of nodes already on the stack
-    set<NodeTraversal> on_stack;
-    // What components did we find? Because of the way strongly connected
-    // components generalizes, both orientations of a node always end up in the
-    // same component.
-    set<set<nid_t> > components;
-    
-    dfs([&](NodeTraversal trav) {
-        // When a NodeTraversal is first visited
-        // It is its own root
-        roots[trav] = trav;
-        // We discovered it at this step
-        discover_idx[trav] = index++;
-        // And it's on the stack
-        stack.push_back(trav);
-        on_stack.insert(trav);
-    },
-        [&](NodeTraversal trav) {
-            // When a NodeTraversal is done being recursed into
-            for (auto next : travs_from(trav)) {
-                // Go through all the NodeTraversals reachable reading onwards from this traversal.
-                if (on_stack.count(next)) {
-                    // If any of those NodeTraversals are on the stack already
-                    auto& node_root = roots[trav];
-                    auto& next_root = roots[next];
-                    // Adopt the root of the NodeTraversal that was discovered first.
-                    roots[trav] = discover_idx[node_root] <
-                    discover_idx[next_root] ?
-                    node_root :
-                    next_root;
-                }
-            }
-            if (roots[trav] == trav) {
-                // If we didn't find a better root
-                NodeTraversal other;
-                set<nid_t> component;
-                do
-                {
-                    // Grab everything that was put on the DFS stack below us
-                    // and put it in our component.
-                    other = stack.back();
-                    stack.pop_back();
-                    on_stack.erase(other);
-                    component.insert(other.node->id());
-                } while (other != trav);
-                components.insert(component);
-            }
-        });
-    
-    return components;
-}
 
 // returns the rank of the node in the protobuf array that backs the graph
 int VG::node_rank(Node* node) {
@@ -3233,71 +2819,6 @@ VG VG::reverse_complement_graph(unordered_map<nid_t, pair<nid_t, bool>>& node_tr
     rev_comp.build_indexes();
     
     return rev_comp;
-}
-
-bool VG::is_acyclic(void) {
-    unordered_set<NodeTraversal> seen;
-    bool acyclic = true;
-    dfs([&](NodeTraversal trav) {
-            // When a node orientation is first visited
-            if (is_self_looping(trav.node)) {
-                acyclic = false;
-            }
-
-            for (auto& next : travs_from(trav)) {
-                if (seen.count(next)) {
-                    acyclic = false;
-                    break;
-                }
-            }
-            if (acyclic) {
-                seen.insert(trav);
-            }
-        },
-        [&](NodeTraversal trav) {
-            // When we leave a node orientation
-
-            // Remove it from the seen array. We may later start from a
-            // different root and see a way into this node in this orientation,
-            // but it's only a cycle if there's a way into this node in this
-            // orientation when it's further up the stack from the node
-            // traversal we are finding the way in from.
-            seen.erase(trav);
-        },
-        [&](void) { // our break function
-            return !acyclic;
-        });
-    return acyclic;
-}
-
-set<set<nid_t> > VG::multinode_strongly_connected_components(void) {
-    set<set<nid_t> > components;
-    for (auto& c : strongly_connected_components()) {
-        if (c.size() > 1) {
-            components.insert(c);
-        }
-    }
-    return components;
-}
-    
-// keeping all components would be redundant, as every node is a self-component
-void VG::keep_multinode_strongly_connected_components(void) {
-    unordered_set<nid_t> keep;
-    for (auto& c : multinode_strongly_connected_components()) {
-        for (auto& id : c) {
-            keep.insert(id);
-        }
-    }
-    unordered_set<Node*> remove;
-    for_each_node([&](Node* n) {
-            if (!keep.count(n->id())) {
-                remove.insert(n);
-            }
-        });
-    for (auto n : remove) {
-        destroy_node(n);
-    }
-    remove_orphan_edges();
 }
 
 size_t VG::size(void) {
@@ -4197,28 +3718,6 @@ Path VG::create_path(const list<NodeTraversal>& nodes) {
         // TODO: Is it OK if we say we're at a mapping at offset 0 of this node, backwards? Or should we offset ourselves to the end?
     }
     return path;
-}
-
-string VG::path_string(const list<NodeTraversal>& nodes) {
-    string seq;
-    for (const NodeTraversal& n : nodes) {
-        if(n.backward) {
-            seq.append(reverse_complement(n.node->sequence()));
-        } else {
-            seq.append(n.node->sequence());
-        }
-    }
-    return seq;
-}
-
-string VG::path_string(const Path& path) {
-    string seq;
-    for (int i = 0; i < path.mapping_size(); ++i) {
-        auto& m = path.mapping(i);
-        Node* n = node_by_id[m.position().node_id()];
-        seq.append(mapping_sequence(m, *n));
-    }
-    return seq;
 }
 
 void VG::expand_path(const list<NodeTraversal>& path, vector<NodeTraversal>& expanded) {
@@ -6439,7 +5938,14 @@ VG VG::dagify(uint32_t expand_scc_steps,
               
     VG dag;
     // Find the strongly connected components in the graph.
-    set<set<nid_t>> strong_components = strongly_connected_components();
+    set<set<nid_t> > strong_components;
+    for (auto& component : handlealgs::strongly_connected_components(this)) {
+        set<nid_t> comp;
+        for (auto node_id : component) {
+            comp.insert(node_id);
+        }
+        strong_components.emplace(move(comp));
+    }
     // map from component root id to a translation
     // that maps the unrolled id to the original node and whether we've inverted or not
 
@@ -6633,464 +6139,6 @@ VG VG::dagify(uint32_t expand_scc_steps,
     // which are valid but can introduce problems for some algorithms
     dag.flip_doubly_reversed_edges();
     return dag;
-}
-// Unrolls the graph into a tree in which loops are "unrolled" into new nodes
-// up to some max length away from the root node and orientations are flipped.
-// A translation between the new nodes that are introduced and the old nodes and graph
-// is returned so that we can map reads against this structure and resolve their mappings back
-// to the original "rolled" and inverted graph.
-// The graph which is returned can be seen as a tree rooted at the source node, so proper
-// selection of the new root may be important for performance.
-// Paths cannot be maintained provided their current implementation.
-// Annotated collections of nodes, or subgraphs, may be a way to preserve the relationshp.
-VG VG::backtracking_unroll(uint32_t max_length, uint32_t max_branch,
-                           unordered_map<nid_t, pair<nid_t, bool> >& node_translation) {
-    VG unrolled;
-    // Find the strongly connected components in the graph.
-    set<set<nid_t>> strong_components = strongly_connected_components();
-    // add in bits where we have inversions that we'd reach from the forward direction
-    // we will "unroll" these regions as well to ensure that all is well
-    // ....
-    //
-    map<nid_t, VG> trees;
-    // maps from entry id to the set of nodes
-    map<nid_t, set<nid_t> > components;
-    // map from component root id to a translation
-    // that maps the unrolled id to the original node and whether we've inverted or not
-    map<nid_t, map<nid_t, pair<nid_t, bool> > > translations;
-    map<nid_t, map<pair<nid_t, bool>, set<nid_t> > > inv_translations;
-
-    // ----------------------------------------------------
-    // unroll the strong components of the graph into trees
-    // ----------------------------------------------------
-    //cerr << "unroll the strong components of the graph into trees" << endl;
-    // Anything outside the strongly connected components can be copied directly into the DAG.
-    // We can also reuse the entry nodes to the strongly connected components.
-    for (auto& component : strong_components) {
-        // is this node a single component?
-        // does this have an inversion as a child?
-        // let's add in inversions
-
-        if (component.size() == 1) {
-            // copy into the new graph
-            nid_t id = *component.begin();
-            node_translation[id] = make_pair(id, false);
-            // we will handle this node if it has an inversion originating from it
-            //if (!has_inverting_edge(get_node(id))) {
-            //nonoverlapping_node_context_without_paths(get_node(id), unrolled);
-            unrolled.add_node(*get_node(id));
-            continue;
-            //}
-            // otherwise we will consider it as a component
-            // and it will be unrolled max_length bp
-        }
-
-        // we have a multi-node component
-        // first find the entry points
-        // entry points will be nodes that have connections outside of the component
-        set<nid_t> entries;
-        set<nid_t> exits;
-        for (auto& n : component) {
-            for (auto& e : edges_of(get_node(n))) {
-                if (!component.count(e->from())) {
-                    entries.insert(n);
-                }
-                if (!component.count(e->to())) {
-                    exits.insert(n);
-                }
-            }
-        }
-
-        // Use backtracking search starting from each entry node of each strongly
-        // connected component, keeping track of the nodes on the path from the
-        // entry node to the current node:
-        for (auto entrypoint : entries) {
-            //cerr << "backtracking search from " << entrypoint << endl;
-            // each entry point is going to make a tree
-            // we can merge them later with some tricks
-            // for now just make the tree
-            VG& tree = trees[entrypoint];
-            components[entrypoint] = component;
-            // maps from new ids to old ones
-            map<nid_t, pair<nid_t, bool> >& trans = translations[entrypoint];
-            // maps from old to new ids
-            map<pair<nid_t, bool>, set<nid_t> >& itrans = inv_translations[entrypoint];
-
-            /*
-            // backtracking search
-
-            procedure bt(c)
-            if reject(P,c) then return
-            if accept(P,c) then output(P,c)
-            s ← first(P,c)
-            while s ≠ Λ do
-            bt(s)
-            s ← next(P,s)
-            */
-
-            function<void(pair<nid_t,bool>,nid_t,bool,uint32_t,uint32_t)>
-                bt = [&](pair<nid_t, bool> curr,
-                         nid_t parent,
-                         bool in_cycle,
-                         uint32_t length,
-                         uint32_t branches) {
-                //cerr << "bt: curr: " << curr.first << " parent: " << parent << " len: " << length << " " << branches << " " << (in_cycle?"loop":"acyc") << endl;
-
-                // i.   If the current node is outside the component,
-                //       terminate this branch and return to the previous branching point.
-                if (!component.count(curr.first)) {
-                    return;
-                }
-                // ii.  Create a new copy of the current node in the DAG
-                //       and use that copy for this branch.
-                Node* node = get_node(curr.first);
-                string curr_node_seq = node->sequence();
-                // if we've reversed, take the reverse complement of the node we're flipping
-                if (curr.second) curr_node_seq = reverse_complement(curr_node_seq);
-                // use the forward orientation by default
-                nid_t cn = tree.create_node(curr_node_seq)->id();
-                // record the mapping from the new id to the old
-                trans[cn] = curr;
-                // and record the inverse mapping
-                itrans[curr].insert(cn);
-                // and build the tree by connecting to our parent
-                if (parent) {
-                    tree.create_edge(parent, cn);
-                }
-
-                // iii. If we have found the first cycle in the current branch
-                //       (the current node is the first one occurring twice in the path),
-                //       initialize path length to 1 for this branch.
-                //       (We need to find a k-path starting from the last offset of the previous node.)
-                // walk the path back to the root to determine if we are the first cycling node
-                nid_t p = cn;
-                // check is borked
-                while (!in_cycle) { // && trans[p] != entrypoint) {
-                    auto parents = tree.sides_to(NodeSide(p, false));
-                    if (parents.size() < 1) { break; }
-                    assert(parents.size() == 1);
-                    p = parents.begin()->node;
-                    // this node cycles
-                    if (trans[p] == trans[cn]) {
-                        in_cycle = true;
-                        //length = 1; // XXX should this be 1? or maybe the length of the node?
-                        break;
-                    }
-                }
-
-                // iv.  If we have found a cycle in the current branch,
-                //       increment path length by the length of the label of the current node.
-                if (in_cycle) {
-                    length += curr_node_seq.length();
-                } else {
-                    // if we branch here so we need to record it
-                    // so as to avoid branching too many times before reaching a loop
-                    auto s = start_degree(node);
-                    auto e = end_degree(node);
-                    branches += max(s-1 + e-1, 0);
-                }
-
-                // v.   If path length >= k, terminate this branch
-                //       and return to the previous branching point.
-                if (length >= max_length || (max_branch > 0 && branches >= max_branch)) {
-                    return;
-                } else {
-                    // for each next node
-                    if (!curr.second) {
-                        // forward direction -- sides from end
-                        for (auto& side : sides_from(node_end(curr.first))) {
-                            // we enter the next node in the forward direction if the side is
-                            // the start, and the reverse if it is the end
-                            bt(make_pair(side.node, side.is_end), cn, in_cycle, length, branches);
-                        }
-                        // this handles the case of doubly-inverted edges (from start to end)
-                        // which we can follow as if they are forward
-                        // we stay on the same strand
-                        for (auto& side : sides_to(node_end(curr.first))) {
-                            // we enter the next node in the forward direction if the side coming
-                            // from the start, and the reverse if it is the end
-                            bt(make_pair(side.node, !side.is_end), cn, in_cycle, length, branches);
-                        }
-                    } else { // we've inverted already
-                        // so we look at the things that leave the node on the reverse strand
-                        // inverted
-                        for (auto& side : sides_from(node_start(curr.first))) {
-                            // here, sides from the start to the end maintain the flip
-                            // but if we go to the start of the next node, we invert back
-                            bt(make_pair(side.node, side.is_end), cn, in_cycle, length, branches);
-                        }
-                        // odd, but important quirk
-                        // we also follow the normal edges, but in the reverse direction
-                        // because we store them in the forward orientation
-                        for (auto& side : sides_to(node_start(curr.first))) {
-                            bt(make_pair(side.node, side.is_end), cn, in_cycle, length, branches);
-                        }
-                    }
-                }
-
-            };
-
-            // we start with the entrypoint and run backtracking search
-            bt(make_pair(entrypoint, false), 0, false, 0, 0);
-        }
-    }
-
-    // -------------------------
-    // tree -> dag conversion
-    // -------------------------
-    // now simplify each tree into a dag
-    // algorithm sketch:
-    // for each tree, we'll make a dag
-    //   1) we start by labeling each node with its rank from the root (we have a tree, then DAGs)
-    //      among nodes with the same original identity
-    //   2) then, we pick the set of nodes that forms the largest group with the same identity
-    //      and merge them
-    //      if there is no group >1, exit, else goto (1), relabeling
-    map<nid_t, VG> dags;
-    for (auto& g : trees) {
-        nid_t entrypoint = g.first;
-        VG& tree = trees[entrypoint];
-        VG& dag = dags[entrypoint];
-        dag = tree; // copy
-        map<nid_t, pair<nid_t, bool> >& trans = translations[entrypoint];
-        map<pair<nid_t, bool>, set<nid_t> >& itrans = inv_translations[entrypoint];
-        // rank among nodes with same original identity labeling procedure
-        map<pair<nid_t, bool>, size_t> orig_off;
-        size_t i = 0;
-        for (auto& j : itrans) {
-            orig_off[j.first] = i;
-            ++i;
-        }
-        // set up the initial vector we'll use when we have no inputs
-        vector<uint32_t> zeros(orig_off.size(), 0);
-        bool stable = false;
-        uint16_t iter = 0;
-        do {
-            // -------------------------
-            // first establish the rank of each node among other nodes with the same original identity
-            // -------------------------
-            // we'll store our positional rank vectors in the current here
-            map<nid_t, vector<uint32_t> > rankmap;
-            // the graph is sorted (and will stay so)
-            // as such we can run across it in sorted order
-            dag.for_each_node([&](Node* n) {
-                    // collect inbound vectors
-                    vector<vector<uint32_t> > iv;
-                    for (auto& side : dag.sides_to(n->id())) {
-                        nid_t in = side.node;
-                        // should be satisfied by partial order property of DAG
-                        assert(rankmap.find(in) != rankmap.end());
-                        assert(!rankmap[in].empty());
-                        iv.push_back(rankmap[in]);
-                    }
-                    // take their maximum
-                    vector<uint32_t> ranks = (iv.empty() ? zeros : vpmax(iv));
-                    // and increment this node so that the rankmap shows our ranks for each
-                    // node in the original set at this point in the graph
-                    ++ranks[orig_off[trans[n->id()]]];
-                    // then save it in the rankmap
-                    rankmap[n->id()] = ranks;
-                });
-            /*
-            for (auto& r : rankmap) {
-                cerr << r.first << ":" << dag.get_node(r.first)->sequence() << " [ ";
-                for (auto c : r.second) {
-                    cerr << c << " ";
-                }
-                cerr << "]" << endl;
-            }
-            */
-
-            // -------------------------
-            // now establish the class relative ranks for each node
-            // -------------------------
-            // maps from node in the current graph to the original identitiy and its rank among
-            // nodes that also are clones of that node
-            map<nid_t, pair<pair<nid_t, bool>, uint32_t> > rank_among_same;
-            dag.for_each_node([&](Node* n) {
-                    nid_t id = n->id();
-                    rank_among_same[id] = make_pair(trans[id], rankmap[id][orig_off[trans[id]]]);
-                });
-            // dump
-            /*
-            for (auto& r : rank_among_same) {
-                cerr << r.first << ":" << dag.get_node(r.first)->sequence()
-                     << " " << r.second.first.first << (r.second.first.second?"-":"+")
-                     << ":" << r.second.second << endl;
-            }
-            */
-            // groups
-            // populate group sizes
-            // groups map from the
-            map<pair<pair<nid_t, bool>, uint32_t>, vector<nid_t> > groups;
-            for (auto& r : rank_among_same) {
-                groups[r.second].push_back(r.first);
-            }
-            // and find the biggest one
-            map<uint16_t, vector<pair<pair<nid_t, bool>, uint32_t> > > groups_by_size;
-            for (auto& g : groups) {
-                groups_by_size[g.second.size()].push_back(g.first);
-            }
-            // do we have a group that can be merged?
-            if (groups_by_size.rbegin()->first > 1) {
-                // -----------------------
-                // merge the nodes that are the same and in the largest group
-                // -----------------------
-                auto orig = groups_by_size.rbegin()->second.front();
-                auto& group = groups[orig];
-                list<Node*> to_merge;
-                for (auto id : group) {
-                    to_merge.push_back(dag.get_node(id));
-                }
-                auto merged = dag.merge_nodes(to_merge);
-                // we've now merged the redundant nodes
-                // now we need to update the translations to reflect the fact
-                // that we no longer have certain nodes in the dag
-                nid_t new_id = merged->id();
-                // remove all old translations from new to old
-                // and insert the new one
-                // do the same for the inverted translations
-                // from old to new
-                set<nid_t>& inv = itrans[orig.first];
-                for (auto id : group) {
-                    trans.erase(id);
-                    inv.erase(id);
-                }
-                // store the new new->old translation
-                trans[new_id] = orig.first;
-                // and its inverse
-                inv.insert(new_id);
-            } else {
-                // we have no group with more than one member
-                // we are done
-                stable = true;
-            }
-            // sort the graph
-            dag.sort();
-        } while (!stable);
-    }
-
-    // recover all the edges that link the nodes in the acyclic components of the graph
-    unrolled.for_each_node([&](Node* n) {
-            // get its edges in the original graph, and check if their endpoints are now
-            // included in the unrolled graph (in which case they must be acyclic)
-            // if they are, add the edge to the graph
-            for (auto e : this->edges_of(get_node(n->id()))) {
-                if (unrolled.has_node(e->from()) && unrolled.has_node(e->to())) {
-                    unrolled.add_edge(*e);
-                }
-            }
-        });
-
-
-    // -----------------------------------------------
-    // connect unrolled components back into the graph
-    // -----------------------------------------------
-    // now that we've unrolled all of the components
-    // what do we do
-    // we link them back into the rest of the graph in two steps
-
-    // for each component we've unrolled (actually, for each entrypoint)
-    for (auto& g : dags) {
-        nid_t entrypoint = g.first;
-        VG& dag = dags[entrypoint];
-        set<nid_t> component = components[entrypoint];
-        map<nid_t, pair<nid_t, bool> >& trans = translations[entrypoint];
-        map<pair<nid_t, bool>, set<nid_t> >& itrans = inv_translations[entrypoint];
-
-        // 1) increment the node ids to not conflict with the rest of the graph
-        //       while recording the changes to the translation
-        nid_t max_id = max_node_id();
-        // update the node ids in the dag
-        dag.increment_node_ids(max_id);
-        map<nid_t, pair<nid_t, bool> > trans_incr;
-        // increment the translation from new node to old
-        for (auto t = trans.begin(); t != trans.end(); ++t) {
-            trans_incr[t->first + max_id] = t->second;
-        }
-        // save the incremented translation
-        trans = trans_incr;
-        // and do the same for the reverse mapping from old ids to new ones
-        for (auto t = itrans.begin(); t != itrans.end(); ++t) {
-            set<nid_t> n;
-            for (auto i = t->second.begin(); i != t->second.end(); ++i) {
-                n.insert(*i + max_id);
-            }
-            t->second = n;
-        }
-        // 2) now that we don't conflict, add the component to the graph
-        unrolled.extend(dag);
-        // and also record the translation into the external reference
-        for (auto t : trans) {
-            // we map from a node id in the graph to an original node id and its orientation
-            node_translation[t.first] = t.second;
-        }
-
-        // 3) find all the links into the component
-        //       then connect all the nodes they
-        //       now relate to using itrans
-        for (auto& i : itrans) {
-            auto old_id = i.first.first;
-            auto is_flipped = i.first.second;
-            set<nid_t>& new_ids = i.second;
-            // collect connections to old id that aren't in the component
-            // we need to take the reverse complement of these if we are flipped relatively
-            // sides to forward
-            // add reverse complement function for edges and nodesides
-            // make the connections
-            for (auto i : new_ids) {
-                // sides to forward
-                for (auto& s : sides_to(NodeSide(old_id, false))) {
-                    // if we are not in the component
-                    if (!component.count(s.node)) {
-                        // if we aren't flipped, we just make the connection
-                        if (!is_flipped) {
-                            unrolled.create_edge(s, NodeSide(i, false));
-                        } else {
-                            // the new node is flipped in unrolled relative to the other graph
-                            // so we need to connect from the opposite side
-                            unrolled.create_edge(s, NodeSide(i, true));
-                        }
-                    }
-                }
-                // sides to reverse
-                for (auto& s : sides_to(NodeSide(old_id, true))) {
-                    // if we are not in the component
-                    if (!component.count(s.node)) {
-                        if (!is_flipped) {
-                            unrolled.create_edge(s, NodeSide(i, true));
-                        } else {
-                            unrolled.create_edge(s, NodeSide(i, false));
-                        }
-                    }
-                }
-                // sides from forward
-                for (auto& s : sides_from(NodeSide(old_id, true))) {
-                    // if we are not in the component
-                    if (!component.count(s.node)) {
-                        if (!is_flipped) {
-                            unrolled.create_edge(NodeSide(i, true), s);
-                        } else {
-                            unrolled.create_edge(NodeSide(i, false), s);
-                        }
-                    }
-                }
-                // sides from reverse
-                for (auto& s : sides_from(NodeSide(old_id, false))) {
-                    // if we are not in the component
-                    if (!component.count(s.node)) {
-                        if (!is_flipped) {
-                            unrolled.create_edge(NodeSide(i, false), s);
-                        } else {
-                            unrolled.create_edge(NodeSide(i, true), s);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    return unrolled;
 }
 
 } // end namespace
